SHIP fixes as part of PR review 9424

+++ libraries/state_history/include/eosio/state_history/serialization.hpp
@@ -313,7 +313,7 @@ ST& operator<<(ST& ds, const history_serial_wrapper<eosio::chain::chain_config>&
    fc::raw::pack(ds, as_type<uint32_t>(obj.obj.max_inline_action_size));
    fc::raw::pack(ds, as_type<uint16_t>(obj.obj.max_inline_action_depth));
    fc::raw::pack(ds, as_type<uint16_t>(obj.obj.max_authority_depth));
-   fc::raw::pack(ds, as_type<uint16_t>(obj.obj.max_action_return_value_size));
+   fc::raw::pack(ds, as_type<uint32_t>(obj.obj.max_action_return_value_size));
    return ds;
 }
